+++ allennlp/modules/transformer/__init__.py
@@ -1,9 +1,8 @@
-from allennlp.modules.transformer.bert_embeddings import BertEmbeddings
+from allennlp.modules.transformer.transformer_embeddings import TransformerEmbeddings
 from allennlp.modules.transformer.self_attention import SelfAttention
-from allennlp.modules.transformer.bert_attention import BertAttention
-from allennlp.modules.transformer.bert_intermediate import BertIntermediate
-from allennlp.modules.transformer.bert_layer import BertLayer
-from allennlp.modules.transformer.biattention import BiAttention
+from allennlp.modules.transformer.activation_layer import ActivationLayer
+from allennlp.modules.transformer.transformer_layer import TransformerLayer
+from allennlp.modules.transformer.bimodal_attention import BiModalAttention
 from allennlp.modules.transformer.output_layer import OutputLayer
 from allennlp.modules.transformer.attention_scores import (
     GeneralAttention,