+++ allennlp/modules/transformer/bimodal_encoder.py
@@ -1,0 +1,7 @@
+
+import torch
+
+from allennlp.common import FromParams
+
+class BiModalEncoder(torch.nn.Module, FromParams):
+	def __init__(self)